Dry up question components
