+++ app/components/questions/true_false_question_component.rb
@@ -1,17 +1,4 @@
-require "ostruct"
-
-class Questions::TrueFalseQuestionComponent < ViewComponent::Base
-  def initialize(question: nil, form: nil, just_answered: false, mock: false)
-    @mock = mock
-    @question = mock ? mock_question : question
-    @form = form
-    @just_answered = just_answered
-  end
-
-  def before_render
-    @just_answered ||= question_answered?
-  end
-
+class Questions::TrueFalseQuestionComponent < Questions::BaseQuestionComponent
   private
 
   def mock_question
@@ -24,13 +11,5 @@ class Questions::TrueFalseQuestionComponent < ViewComponent::Base
 
   def correct_answer
     @question.questionable.correct_answer
-  end
-
-  def question_answered?
-    return false if @mock
-    @question.user_answers.where(
-      user: current_user,
-      progression: @question.course.progressions.where(user: current_user).order(id: :desc).first
-    ).exists?
   end
 end
